ユーティリティの整理
- クリップボード周りにpyperclip導入した

- ブラウザ開くのにurlによってミスるからやりかた変えた
