+++ pyproject.toml
@@ -12,6 +12,7 @@ dependencies = [
     "numpy>=2.1.2",
     "opencv-python>=4.10.0.84",
     "pillow>=10.4.0",
+    "pyperclip>=1.9.0",
     "qrcode>=8.0",
 ]
